+++ pyoxidizer/src/python_distributions.rs
@@ -20,9 +20,9 @@ lazy_static! {
         res.insert(
             "x86_64-unknown-linux-gnu".to_string(),
             HostedDistribution {
-                url: String::from("https://github.com/indygreg/python-build-standalone/releases/download/20200216/cpython-3.7.6-linux64-20200216T2303.tar.zst"),
+                url: String::from("https://github.com/indygreg/python-build-standalone/releases/download/20200408/cpython-3.7.7-linux64-20200409T0045.tar.zst"),
                 sha256: String::from(
-                    "58067eecbd1600ea765f7fe7b43562bbc8058db4c84ddbfcaddcd2ee18193907",
+                    "74799ae3b7f3ddc2d118516d65d46356fb3ef3ff3c4c4591a0dde073c413aff0",
                 ),
             },
         );
@@ -30,9 +30,9 @@ lazy_static! {
         res.insert(
             "x86_64-unknown-linux-musl".to_string(),
             HostedDistribution {
-                url: String::from("https://github.com/indygreg/python-build-standalone/releases/download/20200217/cpython-3.7.6-linux64-musl-20200218T0557.tar.zst"),
+                url: String::from("https://github.com/indygreg/python-build-standalone/releases/download/20200408/cpython-3.7.7-linux64-musl-20200409T0047.tar.zst"),
                 sha256: String::from(
-                    "d5e5d8681b0af13bc3e718a35d6237b3629908a669050cb8c8ab919a731c5718",
+                    "c1ffa330c7305f46886b7cd2b77edf0e43463113cef426d388476337c3e5cfa9",
                 ),
             },
         );
@@ -40,24 +40,24 @@ lazy_static! {
         res.insert(
             "i686-pc-windows-msvc".to_string(),
             HostedDistribution {
-                url: String::from("https://github.com/indygreg/python-build-standalone/releases/download/20200216/cpython-3.7.6-windows-x86-static-20200216T2309.tar.zst"),
-                sha256: String::from("29fcca67a022bfac3f29a8a32cb070eef8cecbf052dd9a4eff6feae441ca8fb6"),
+                url: String::from("https://github.com/indygreg/python-build-standalone/releases/download/20200408/cpython-3.7.7-windows-x86-static-20200409T0107.tar.zst"),
+                sha256: String::from("978e863fd39f8758c2af18dd750f64eb4b57bd8bfea86cae5fcdde305c56dca7"),
             },
         );
 
         res.insert(
             "x86_64-pc-windows-msvc".to_string(),
             HostedDistribution {
-                url: String::from("https://github.com/indygreg/python-build-standalone/releases/download/20200216/cpython-3.7.6-windows-amd64-static-20200216T2300.tar.zst"),
-                sha256: String::from("a9348f50d7289fde92e0ad36073febf9b86448737286cecead08881d009b5829"),
+                url: String::from("https://github.com/indygreg/python-build-standalone/releases/download/20200408/cpython-3.7.7-windows-amd64-static-20200409T0105.tar.zst"),
+                sha256: String::from("fe8d95bc2d7d911ba23c318b786ea7d17c3e2aadbedf47b1d53962bf42e418fe"),
             },
         );
 
         res.insert(
             "x86_64-apple-darwin".to_string(),
             HostedDistribution {
-                url: String::from("https://github.com/indygreg/python-build-standalone/releases/download/20200216/cpython-3.7.6-macos-20200216T2344.tar.zst"),
-                sha256: String::from("0487f70c2b857ddcf8d005ba6de9b97b13800aeaeb505c42da7659d83c79d233"),
+                url: String::from("https://github.com/indygreg/python-build-standalone/releases/download/20200408/cpython-3.7.7-macos-20200409T0412.tar.zst"),
+                sha256: String::from("f312bea46a7d8efecd4df6b22c03f83016775e6bb5944a5701d697e0a52c63b2"),
             },
         );
 
@@ -69,16 +69,16 @@ lazy_static! {
         res.insert(
             "i686-pc-windows-msvc".to_string(),
             HostedDistribution {
-                url: String::from("https://github.com/indygreg/python-build-standalone/releases/download/20200216/cpython-3.7.6-windows-x86-shared-pgo-20200217T0110.tar.zst"),
-                sha256: String::from("a77b2245f0109fa80cd46adeb40815a1e8892002fffa64293a9702f50d547bc2"),
+                url: String::from("https://github.com/indygreg/python-build-standalone/releases/download/20200408/cpython-3.7.7-windows-x86-shared-pgo-20200409T0157.tar.zst"),
+                sha256: String::from("69813ae54e691e244e02b25099f7af90a0bac1b63eae80dfe039b7f26010072e"),
             },
         );
 
         res.insert(
             "x86_64-pc-windows-msvc".to_string(),
             HostedDistribution {
-                url: String::from("https://github.com/indygreg/python-build-standalone/releases/download/20200216/cpython-3.7.6-windows-amd64-shared-pgo-20200217T0022.tar.zst"),
-                sha256: String::from("35ccece4950147a9344e4843bc6148882b12b79806707726b15e846eb6cfed4e"),
+                url: String::from("https://github.com/indygreg/python-build-standalone/releases/download/20200408/cpython-3.7.7-windows-amd64-shared-pgo-20200409T0115.tar.zst"),
+                sha256: String::from("c43c44ebfe9b9f9c59c12481a6233b17bc9a4ad965f8f0dc0063abff4dc59875"),
             },
         );
 
@@ -90,9 +90,9 @@ lazy_static! {
         res.insert(
             "i686-pc-windows-msvc".to_string(),
             HostedDistribution {
-                url: "https://www.python.org/ftp/python/3.7.6/python-3.7.6-embed-win32.zip"
+                url: "https://www.python.org/ftp/python/3.7.7/python-3.7.7-embed-win32.zip"
                     .to_string(),
-                sha256: "e2257b87e2e1a131e5d2adf843887fdab5021f8d4d6d68d49691aa965650c3ab"
+                sha256: "0f3742ec2eeefa426f818997df9aff578a73e424e836b6fb38cb1bb243d63d31"
                     .to_string(),
             },
         );
@@ -100,9 +100,9 @@ lazy_static! {
         res.insert(
             "x86_64-pc-windows-msvc".to_string(),
             HostedDistribution {
-                url: "https://www.python.org/ftp/python/3.7.6/python-3.7.6-embed-amd64.zip"
+                url: "https://www.python.org/ftp/python/3.7.7/python-3.7.7-embed-amd64.zip"
                     .to_string(),
-                sha256: "114638061d636285600cbc3d4def64b45c43da9b225cb9eeead30fe7fe7d60d4"
+                sha256: "705c03140cfd3372f27ee911db5f4eb1fc9b980c9e27544adbd1a6adf942a1b0"
                     .to_string(),
             },
         );